Database created
Database for storing playlists has been created and finished. It still needs testing

+++ app/src/main/java/Song.kt
@@ -8,7 +8,7 @@ class Song(
     val path: String, val fileSize: Int,
     val lyricsPath: String?) {
 
-    constructor(song: SongEntity) : this(song.title, song.artist, song.album, song.duration,
+    constructor(song: SongEntity) : this(song.title, song.artist, song.album, song.duration.toUInt(),
                                          song.path, song.size, song.lyricsPath)
     fun getSizeMB(): Float{
         //1024 * 1024 = 1048567
@@ -21,5 +21,5 @@ class Song(
         return df.format(fileSize.toFloat() / TAMANO_MB).toFloat()
     }
 
-    fun toSongEntity() = SongEntity(path, title, artist, album, duration, fileSize, lyricsPath)
+    fun toSongEntity() = SongEntity(path, title, artist, album, duration.toInt(), fileSize, lyricsPath)
 }
